compositor|tests: Share one overlapping-target builder across the plan suites
Moves the duplicated target builder into `plan/test-utils/composition-fixture.ts`, the nearest common ancestor of the two suites that reach it and the module both already import `buildClaudeTarget` from.

Names it `buildOverlappingTargets`, after the configuration rather than either outcome it produces. The two deployments are rooted at one directory, and that overlap is what makes a claimed path's provenance undecidable in one test and a planned destination contested in two others, so neither `ambiguous` nor `contested` describes what the builder builds.
